video: mcde: Add console semaphore
Earlysuspend and Lateresume has to use the console semphore.
This semaphore is taken in fbmem.c for every call to mcde_fb.

If this semaphore is not acquired in earlysuspend, a suspend can
occur at the same time as a refresh is ongoing.
This can result in unbalanced disables for VAUX1.

Signed-off-by: Jimmy Rubin <jimmy.rubin@stericsson.com>