serial/amba-pl011: prevent console port from holding regulator
Register the console in a console_initcall() to ensure that
pl011_console_setup() gets called early and thus the
uart_console() macro usage in pl011_power_startup() returns
the correct value the first time it is called, so that the
vAPE regulator is not taken for the console port.

Note that this problem is seen only when the console port
is line 0, since even without this patch pl011_console_setup()
gets called (and thus the correct console->index is set)
after the first port is added.

ST-Ericsson ID: 368522
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I7f4ab80e8d8be014bfdbe695ea45007212c9c4ea
Signed-off-by: Rabin Vincent <rabin.vincent@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/34411
Reviewed-by: Srinidhi KASAGAR <srinidhi.kasagar@stericsson.com>
Reviewed-by: Shreshtha Kumar SAHU <shreshthakumar.sahu@stericsson.com>