ux500: ARM: u8500: power: idle test
Add power module test. This test, in debugfs/pwr_test/idle shall
pass when the screen is black, but suspend is not yet reached.
This module test checks clocks, prcmu-regulators, db8500 regulators
and ab8500 regulators.

The reason why not adding this to LTP is that
later this module test will contain tests for voicecall and
low-power-audio, which require Android to run.

ST-Ericsson ID: 367599

Signed-off-by: Jonas Aaberg <jonas.aberg@stericsson.com>