ux500: TEE: Moved conversion of virt_to_phys
If memory allocation failed when trying to copy memrefs from user space
to kernel space we jumped to error handling code which tried to free the
memrefs that had been allocated previously. The problem was that after
each allocation we converted the allocated buffers to physical
addresses, hence if any of the allocations except the first one failed
we ended up trying to free a physical address. In this patch we have
moved the translation from virtual to physical addresses to be done
after all allocations have succeeded.

ST-Ericsson ID: 374810, 374493, 374920
ST-Ericsson FOSS-OUT ID: NA
ST-Ericsson Linux next: NA

Change-Id: I3a109a9ebb46f74a2089916fa300bf6f4347501b
Signed-off-by: Joakim Bech <joakim.xx.bech@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/39876
Reviewed-by: QATOOLS
Reviewed-by: QABUILD
Reviewed-by: QATEST
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>