u5500: cw1200: Remove requesting SDIO pins
The SDIO pins belongs to the SDIO driver and
should be handled by the SDIO driver only.

The reason that it has worked up until now is that
the SDIO driver has just started using pm runtime.
Since both WLAN and SDIO framework enable
the pins, the number of users is set to two.

SDIO tries to release the pins (after 50 ms of inactivity
in order to save power), the usage count does not go
down to 0 and therefore the pins are not reconfigured
to GPIOs, which they must be when PL18X is not driving
the pins due to power down).

ST-Ericsson ID: 375493
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: NA

Change-Id: Ibe3dfd32cb16a9d629a8c405e5723aec787a5ae6
Signed-off-by: Bartosz Markowski <bartosz.markowski@tieto.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/41183
Reviewed-by: QATOOLS
Reviewed-by: QABUILD
Reviewed-by: Srinidhi KASAGAR <srinidhi.kasagar@stericsson.com>