caif-shm: synchronize caif flow control
In the shm tx work function caif flow control is turned ON and without
acquiring the spin lock. In the rx callback cail flow control is turned
off. The expected sequence would be on caif writing message to modem
via shm, if shm finds that there is no empty space then, it is not sets
caif flow control to OFF. On shm receiving messages from modem, it sets
caif flow control to ON. Now this since this is in interrupt context, there
are chances of this fucntion being preeempted by the caif OFF that is being
set in the tx work function.
Hence create a seperate workqueue and two work functions to turn caif flow
control ON and OFF. With this implementation it is ensured that caif flow
control is turned OFF first by tx work function and turned ON by the rx
callback.

ST-Ericsson Linux next: NA
ST-Ericsson ID: 372652
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ide75415233d7b4560faf25862876c15421df7c2c
Signed-off-by: Arun Murthy <arun.murthy@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/42438
Reviewed-by: QABUILD
Reviewed-by: Hemant-vilas RAMDASI <hemant.ramdasi@stericsson.com>
Reviewed-by: Sjur BRENDELAND <sjur.brandeland@stericsson.com>