u5500:Ignored link interrupt not to post usb event
The ABB link status interrupt is coming twice for
LinkStatus = 18 and 13 ,one of them is to be
ignored depending on the current ABB hardware
version. However for the ignored one too, the ABB
driver is sending a spurious USB_EVENT_NONE notification
event ,which is thus shutting down the vbus needed
for enumerating the device. Hence 5500 usb host is
not enumerating any devices.

Solution is not to post the event USB_EVENT_NONE
for the ignored link status interrupt.

ST-Ericsson ID: 401855
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I7a7d8a19a5431e21661a0875061d51ba0129f25e
Signed-off-by: Ravikant Singh <ravikant.singh@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/43042
Reviewed-by: Praveena NADAHALLY <praveen.nadahally@stericsson.com>
Reviewed-by: Rabin VINCENT <rabin.vincent@stericsson.com>