cg2900: [U9500] Correct the GPIO configuration for WLAN
There are two parts in this correction.
Firstly, as there is much in common between a U9500 platform
and a U8500 one, all the tests performed to identify a U9500 platform
shall be done before the ones related to U8500. Otherwise,
the platform will be considered as a U8500 one and the GPIOs will
be badly configured.
Secondly, the mapping of WLAN_IRQ has to be changed. On U8500, WLAN_IRQ
is mapped onto GPIO04. On U9500, it is GPIO144 that shall be used.

ST-Ericsson ID: 368270
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial

Signed-off-by: Rejane DURAND <rejane.durand@stericsson.com>