b2r2: Possible to set YUV range
The YUV range can be either full or narrow (601 standard).
A flag has been added to be able to choose which YUV range
to use for the B2R2 request. The conversion matrices (iVMx)
has also been updated to default to narrow range. Narrow range
is the preferred range since it is used for video encoding.
The full range case is targeted for JPEG encoding.

Conversion matrix assignment has been broken out to a
separate file. As a consequence of this some functionality
in the node splitter has been moved to utilis.

ST-Ericsson ID: 407107
ST-Ericsson FOSS-OUT ID: Trivial
ST-Ericsson Linux next: NA

Change-Id: Id4edb38170b0c398927ad4eddbdecc62e4734cdf
Signed-off-by: Jorgen Nilsson <jorgen.nilsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/46251
Reviewed-by: Jimmy RUBIN <jimmy.rubin@stericsson.com>
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>