cpufreq: ondemand: handle SW coordinated CPUs
This patch fixes a bug that occured when we had load on
a secondary CPU and the primary CPU was sleeping. Only
one sampling timer was spawned and it was spawned as a
deferred timer on the primary CPU, so when a secondary CPU
had a change in load this was not detected by the ondemand
governor.

This patch make sure that deferred timers are run on all
CPUs in the case of software controlled CPUs that run on
the same frequency.

Signed-off-by: Rickard Andersson <rickard.andersson@stericsson.com>
Reviewed-by: Mian Yousaf Kaukab <mian.yousaf.kaukab@stericsson.com>
Reviewed-by: Jonas Aberg <jonas.aberg@stericsson.com>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
---
ChangeLog v1->v2: forgot to include the refactorings for
 cpu device rewrites, which is the version I actually tested,
 mea culpa.