cw1200: block consecutive calls to wsm_set_pm
Reported by Dheeshan SASI:
One observation in OMAC logging is that there are repeated calls
of set-PM-Mode-Req. This is incorrect, it is a call that should be
made only once. After this, the FW manages power control. Each time
the call is made the FW states have to be reset.

Even on channel changes (as in code snippet below) FW will manage
the power save states. Host MAC should not control this. Leave always
in fast power save, the only practical reason to change this is for
rogue/bad APs. Channel change will internally change the joined channel.

If there appears to be a link to Power Save, then this set-PM-Mode-Req
issue should be corrected. The call can be made once per association,
after set-bss-params. This means all the information for power save is
available (AID especially). Also after Reset if necessary for OMAC.

Similarly operational mode also should only be called once. There is
no need to keep calling this, however this has less impact and no large
state machine is involved.

ST-Ericsson ID: 355071

Change-Id: I740e9b21f5e1fdf0b696ddc709b26a640872922c
Signed-off-by: Dmitry Tarnyagin <dmitry.tarnyagin@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/37688
Reviewed-by: Bartosz MARKOWSKI <bartosz.markowski@tieto.com>
Tested-by: Bartosz MARKOWSKI <bartosz.markowski@tieto.com>