cw1200: Workaround against throughput drop in b/g mixed mode.
Device has problems (at least) switching from g-rates CTS to 1Mbps.
This switch takes enormous amount of time (100-200 ms), leading to
valuable throughput drop.

As a workaround, additional g-rates are injected to the TX
rate policy if initial rate is one of high g-rates and the
fallback rate is a b-rate.

ST-Ericsson ID: 402230

Change-Id: Ie90195035e77252310e4d9edc0f06b2a86b39712
Signed-off-by: Dmitry Tarnyagin <dmitry.tarnyagin@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/42418
Tested-by: Bartosz MARKOWSKI <bartosz.markowski@tieto.com>
Reviewed-by: Bartosz MARKOWSKI <bartosz.markowski@tieto.com>