usb: musb: SW workaround for USB host issue
For snowball in HOST mode, the eye diagram is facing high distortion
in high speed use cases. This patch allows the user to use an
external charge pump instead of the internal one to generate
the Vbus voltage. This can be done by setting a sysfs entry.

The external VBUS is OFF by default and can be made ON by setting
the sysfs entry inorder to enumerate high speed devices.

Signed-off-by: supriya karanth <supriya.karanth@stericsson.com>